functional main added

+++ JumpDistribution.cpp
@@ -17,7 +17,8 @@ range[0] = r1;
 range[1] = r2;
 MonteCarloObjectPtr = MC;
 
-
+/* Check that the values input externally produce a histogram
+ * and throw exception if not */
 if (range[0] > range[1] || bins < 1 || interval < 0){
     cout << "Histogram bins not well defined" << endl << endl;
     throw "Histgram bins not well defined!";
@@ -44,20 +45,26 @@ void JumpDistribution::makeDistribution(){
 double time = interval;
 int jump = 0;
 
+/* Loop over data until either the interval moves past
+ * the last time, or the last data point is reached */
 while (time < MonteCarloObjectPtr->jumpTimes.back() &&
-    jump < MonteCarloObjectPtr->jumpTimes.size()){
+    jump < static_cast<int>( MonteCarloObjectPtr->jumpTimes.size() ) ){
 
     int jumps = 0;
 
+    //Sum jumps in time interval
     while (MonteCarloObjectPtr->jumpTimes[jump] < time){
             jumps += 1;
             jump += 1;
     }
 
+    //Move to next time interval
     time += interval;
 
+    //Locate the bin...
     int bin = static_cast<int> (jumps - range[0])/binWidth;
 
+    //...and add one to the samples for that bin
     if (bin < bins && bin >=0){
         frequency[bin] += 1;
     }
